backend/requirements.txt
fastapi
uvicorn[standard]
pymongo[srv]
python-dotenv
pandas
openpyxl
requests